+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- CYGNSS soil moisture reader
 - New update_type for joint 3d soil moisture and 1d snow analysis (Tb+sfmc+sfds+SCF obs)
 
 ### Changed